update.sh : add support for other systems

+++ debian/trixie/arm64/update.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+if [ "$1" == "" ]; then
+  echo usage : update.sh {system}
+  exit
+fi
+
+if [ ! -f $1.list.input ]; then
+  echo That system was not found!
+  exit
+fi
+
 sudo apt --yes install dpkg-dev apt-utils
 
 rm *.gz 2>/dev/null
@@ -7,8 +17,8 @@ rm InRelease 2>/dev/null
 rm Release 2>/dev/null
 rm Release.gpg 2>/dev/null
 
-cp ~/.gnupg/pubring.gpg ./javaforce.gpg
-chmod 644 javaforce.gpg
+cp ~/.gnupg/pubring.gpg ./$1.gpg
+chmod 644 $1.gpg
 
 dpkg-scanpackages . > Packages
 
@@ -25,6 +35,6 @@ gpg -abs -o Release.gpg Release
 #gzip Packages
 
 . /etc/os-release
-sed 's/$(VERSION)/'$VERSION_CODENAME'/g' < javaforce.list.input > javaforce.list
+sed 's/$(VERSION)/'$VERSION_CODENAME'/g' < $1.list.input > $1.list
 
 echo Update complete!
